Add adminApi policy to the delete endpoint
