+++ backend/src/StamAcasa.Api/Controllers/ProfileController.cs
@@ -166,21 +166,5 @@ namespace StamAcasa.Api.Controllers
 
             return new OkObjectResult(result != null);
         }
-
-        /// <summary>
-        /// Delete user profile and related data.
-        /// </summary>
-        /// <returns></returns>
-        [HttpDelete]
-        public async Task<IActionResult> DeleteProfile()
-        {
-            var sub = User.Claims.FirstOrDefault(c => c.Type == "sub")?.Value;
-            if (sub == null)
-                return new UnauthorizedResult();
-
-            await _userService.DeleteUserAndDependentData(sub);
-                        
-            return Ok();
-        }
     }
 }
