+++ backend/src/StamAcasa.Api/appsettings.json
@@ -19,8 +19,12 @@
       "ApiSecret": "svpqYnJSR8xzn8Rl"
     },
     {
-      "ApiName": "usersApi",
-      "ApiSecret": "st4k!b7s$af201cv"
+        "ApiName": "usersApi",
+        "ApiSecret": "st4k!b7s$af201cv"
+    },
+    {
+        "ApiName":  "adminApi",
+        "ApiSecret": "3rXGcXpcPKxAIshJ"
     }
   ],
   "TargetFolder": "answers",
