+++ backend/src/StamAcasa.IdentityServer/appsettings.json
@@ -20,6 +20,11 @@
             "Name": "usersApi",
             "Secret": "st4k!b7s$af201cv",
             "ClaimList": [ "openid", "email" ]
+        },
+        {
+            "Name": "adminApi",
+            "Secret": "3rXGcXpcPKxAIshJ",
+            "ClaimList": [ "openid" ]
         }
     ],
     "ClientApplications": [
@@ -124,6 +129,21 @@
                 "usersApi"
             ],
             "AllowAccessTokensViaBrowser": true,
+            "AccessTokenType": 1
+        },
+        {
+            "ClientId": "idsrvClient",
+            "ClientSecrets": [
+                {
+                    "Value": "HcBFXeMlI/5iUCtyLErjz4rJ66u2Bdk+ytqj0aELxhw="
+                }
+            ],
+            "ClientName": "Identiy Server Client",
+            "AllowedGrantTypes": [ "client_credentials" ],
+            "AllowedScopes": [
+                "openid",
+                "adminApi"
+            ],
             "AccessTokenType": 1
         }
     ],
